+++ tests/testthat/test-trajectory.R
@@ -5,6 +5,7 @@ t0 <- create_trajectory(verbose=TRUE) %>%
   select(c("a", "b")) %>%
   seize_selected(1) %>%
   timeout(function() rnorm(1, 15)) %>%
+  leave(0) %>%
   branch(function() 1, T, create_trajectory(verbose=TRUE)%>%timeout(function() 1)) %>%
   set_attribute("dummy", 1) %>%
   rollback(1) %>%
@@ -15,6 +16,7 @@ trajs <- c(create_trajectory(verbose=TRUE) %>% seize("nurse", 1),
            create_trajectory(verbose=TRUE) %>% select(c("a", "b")),
            create_trajectory(verbose=TRUE) %>% seize_selected(1),
            create_trajectory(verbose=TRUE) %>% timeout(function() rnorm(1, 15)),
+           create_trajectory(verbose=TRUE) %>% leave(0),
            create_trajectory(verbose=TRUE) %>% branch(function() 1, T, create_trajectory(verbose=TRUE)%>%timeout(function() 1)),
            create_trajectory(verbose=TRUE) %>% set_attribute("dummy", 1),
            create_trajectory(verbose=TRUE) %>% rollback(1),
@@ -23,9 +25,9 @@ trajs <- c(create_trajectory(verbose=TRUE) %>% seize("nurse", 1),
 
 test_that("the activity chain grows as expected", {
   head <- t0%>%get_head()
-  for (i in 1:8) head <- get_next_activity(head)
+  for (i in 1:9) head <- get_next_activity(head)
   tail <- t0%>%get_tail()
-  for (i in 1:8) tail <- get_prev_activity(tail)
+  for (i in 1:9) tail <- get_prev_activity(tail)
   
   expect_output(print_activity(head), "Release")
   expect_output(print_activity(t0%>%get_tail()), "Release")
@@ -39,9 +41,9 @@ test_that("the activity chain grows as expected using join", {
   t <- join(trajs)
   
   head <- t%>%get_head()
-  for (i in 1:8) head <- get_next_activity(head)
+  for (i in 1:9) head <- get_next_activity(head)
   tail <- t%>%get_tail()
-  for (i in 1:8) tail <- get_prev_activity(tail)
+  for (i in 1:9) tail <- get_prev_activity(tail)
   
   expect_output(print_activity(head), "Release")
   expect_output(print_activity(t%>%get_tail()), "Release")